Create SkyD2H.java
Well! This Project is Developed for SkyD2H. This project is fully developed by OOPS concept in Java and we using Collection framework to store the Customer details in ArrayList and iterate from the ArrayList.This will declare the cost of the package, the number of the customer gets connected and Cost of various packages.